requirements.txt
Flask>=3.0.0,<4.0
Flask-SocketIO>=5.3.6,<6.0
python-socketio>=5.10.0,<6.0
python-engineio>=4.8.1,<5.0
Werkzeug>=3.0.1,<4.0
python-dotenv>=1.0.0
APScheduler>=3.10.0,<4.0
